+++ pages/page-studio-dei-sutra.php
@@ -1,3 +1,7 @@
+<article>
+<h1 class="wp-block-heading">Studio dei Sutra</h1>
+
 <div class="cz-sutra-memo-container">
 <?php echo do_shortcode('[cz-sutra-memo]'); ?>
 </div>
+</article>