+++ CHANGELOG.md
@@ -5,10 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.7.3]
 ### Added
 - Add the page `genealogia-dei-maestri`
-- Add the page `memorizzazione-sutra`
+- Add the page `studio-dei-sutra`
 - Add the page `utilita`
 ### Changed
 - Updates `footer.php` adding links to Instagram and the page "Utilità".
@@ -264,7 +264,8 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v1.7.2...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v1.7.3...HEAD
+[1.7.3]: https://github.com/erremauro/cigno-zen/releases/tag/v1.7.3
 [1.7.2]: https://github.com/erremauro/cigno-zen/releases/tag/v1.7.2
 [1.7.1]: https://github.com/erremauro/cigno-zen/releases/tag/v1.7.1
 [1.7.0]: https://github.com/erremauro/cigno-zen/releases/tag/v1.7.0